refactor: update patch for authApi
